Move WaypointInterface
